==> 01-cleaning-data.ipynb <==
import pandas as pd
import re
import warnings
warnings.filterwarnings('ignore')

Xtr = pd.read_csv('Xtr.csv')
Xte = pd.read_csv('Xte.csv')
Xtr = Xtr[['ID','Address']]
Xtr['Address'] = Xtr['Address'].str.replace('[^a-zA-Z0-9\s]', '', regex=True)
Xte['Address'] = Xte['Address'].str.replace('[^a-zA-Z0-9\s]', '', regex=True)
Xtr['Address'] = Xtr['Address'].str.lstrip()
Xte['Address'] = Xte['Address'].str.lstrip()
Xtr['Address'] = Xtr['Address'].str.replace(r"\s+[A-Z]$", "", regex=True)
Xte['Address'] = Xte['Address'].str.replace(r"\s+[A-Z]$", "", regex=True)
Xtr['Address'] = Xtr['Address'].str.replace(r'\s+', ' ', regex=True)
Xte['Address'] = Xte['Address'].str.replace(r'\s+', ' ', regex=True)

abbreviation_mapping = {
    'VLG': 'VILLAGE',
    'VILL': 'VILLAGE',
    'ST': 'STREET',
    'RD': 'ROAD',
    'LN': 'LANE',
    'LA' : 'LANE',
    'AVE': 'AVENUE',
    'AV': 'AVENUE',
    'RDG': 'RIDGE',
    'BLVD': 'BOULEVARD',
    'RTE': 'ROUTE',
    'UT': 'UNIT',
    'CIR': 'CIRCLE',
    'ISLE': 'IS',
    'CT' : 'COURT',
    'DR' : 'DRIVE',
    'PL' : 'PLACE',
    'HL' : 'HILL',
    'PT' : 'POINT',
    'TR' : 'TERRACE',
    'TER' : 'TERRACE',
    'TERR' : 'TERRACE',
    'HWY' : 'HIGHWAY',
    'HGHY' : 'HIGHWAY',
    'HGWY' : 'HIGHWAY',
    'TRL' : 'TRAIL',
    'SQ' : 'SQUARE',
    'LT' : 'LOT'
}

direction_mapping = {
    'W ': 'WEST ',
    'E ': 'EAST ',
    'N ': 'NORTH ',
    'S ': 'SOUTH '
}
pattern = r'^(' + '|'.join(re.escape(key) for key in direction_mapping.keys()) + r')'



def replace_directions(match):
    return direction_mapping[match.group(0)]
Xtr['Address'] = Xtr['Address'].str.replace(pattern, replace_directions, regex=True)
Xte['Address'] = Xte['Address'].str.replace(pattern, replace_directions, regex=True)
Xtr['Address'] = Xtr['Address'].str.replace(pattern, r'\1', regex=True)
Xte['Address'] = Xte['Address'].str.replace(pattern, r'\1', regex=True)
pattern = r'^[A-C]\s'
Xtr['Address'] = Xtr['Address'].str.replace(pattern, '', regex=True)
Xte['Address'] = Xte['Address'].str.replace(pattern, '', regex=True)
pattern = r'^[A-C]\s'
Xtr['Address'] = Xtr['Address'].str.replace(pattern, '', regex=True)
Xte['Address'] = Xte['Address'].str.replace(pattern, '', regex=True)



patterns_and_replacements = [
    (r'\b(NORTH LAKE DRIVE)\b', r'NORTH LAKE DRIVE'),
    (r'\b(WOLFPIT AVENUE)\b', r'WOLFPIT AVENUE'),
    (r'\b(DONAHUE ROAD)\b', r'DONAHUE ROAD'),
    (r'\b(WEST STREET)\b', r'WEST STREET'),
    (r'\b(SHEPARD AVENUE)\b', r'SHEPARD AVENUE'),
    (r'\b(ROWAYTON WOODS DRIVE)\b', r'ROWAYTON WOODS DRIVE'),
    (r'\b(RIVERTON ROADUNIT)\b', r'RIVERTON ROADUNIT'),
    (r'\b(REMINGTON STREET)\b', r'REMINGTON STREET'),
    (r'\b(PUTNAM HILL)\b', r'PUTNAM HILL'),
    (r'\b(YANTIC STREET)\b', r'YANTIC STREET'),
    (r'\b(VICTORY STREET)\b', r'VICTORY STREET'),
    (r'\b(LAKESIDE BOULEVARD)\b', r'LAKESIDE BOULEVARD'),
    (r'\b(HIGH STREET)\b', r'HIGH STREET'),
    (r'\b(AMSTERDAM AVENUE)\b', r'AMSTERDAM AVENUE'),
    (r'\b(ABBOTT ROAD)\b', r'ABBOTT ROAD'),
    (r'\b(AIKEN STREET)\b', r'AIKEN STREET'),
    (r'\b(ANN STREET)\b', r'ANN STREET'),
    (r'\b(ATLANTIC STREET)\b', r'ATLANTIC STREET'),
    (r'\b(BETHS AVENUE)\b', r'BETHS AVENUE'),
    
    # Add more patterns and replacements here
]

for pattern, replacement in patterns_and_replacements:
    Xtr['Address'] = Xtr['Address'].str.replace(pattern, replacement)
    Xte['Address'] = Xte['Address'].str.replace(pattern, replacement)

Xtr.to_csv('Xtr1.csv', index=False)
Xte.to_csv('Xte1.csv', index=False)
