==> 02.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')
Xtr = pd.read_csv('Xtr.csv')
Xte = pd.read_csv('Xte.csv')
Ytr = pd.read_csv('Ytr.csv')
Xte.loc[Xte['Town'] == '***Unknown***', 'Town'] = 'East Hampton'
Xtr2 = pd.read_csv('Xtr1.csv')
Xte2 = pd.read_csv('Xte1.csv')

Xtr['Address'] = pd.merge(Xtr, Xtr2[['ID', 'Address']], on='ID', how='left')['Address_y']
Xte['Address'] = pd.merge(Xte, Xte2[['ID', 'Address']], on='ID', how='left')['Address_y']
Xtr1 = Xtr.copy()
Xte1 = Xte.copy()
Xtr1['Address'] = Xtr1.groupby('Town')['Address'].transform(lambda x: x.fillna(x.mode()[0]))
Xte1['Address'] = Xte1.groupby('Town')['Address'].transform(lambda x: x.fillna(x.mode()[0]))
Xtr1['Assessed_Value'] = Xtr1.groupby(['Address'])['Assessed_Value'].transform(lambda x: x.fillna(x.median()))
Xte1['Assessed_Value'] = Xte1.groupby(['Address'])['Assessed_Value'].transform(lambda x: x.fillna(x.median()))
Xtr1.to_csv('Xtr_2.csv', index=False)
Xte1.to_csv('Xte_2.csv', index=False)
